requirements.txt
viam-sdk  >= 0.5.1
deepface
pillow
tensorflow==2.14.0rc1